优化了函数的类型名和文件注释，代码简洁如风~

+++ lib/ESP32_menu_v1/1_driver/keyboard/user_defined.h
@@ -4,7 +4,7 @@
 //把你自己的单次键盘扫描函数放在这里                 |
 //这里的函数由你自己调用，if你不用也可以空着         |
 //只要你的扫描-传值函数正常工作就行了                ↓
-int default_scan_once_function();
+uint8_t default_scan_once_function();
 
 //把你自己的完整键盘 扫描-传值 函数放在这里  |
 //这里的函数会定时被timer中断执行           ↓
